edemco-backend-python/microservice_upload-file/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
pandas==2.2.2
python-dateutil==2.9.0